backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
GitPython==3.1.43
langchain-openai==0.1.8
langchain-core==0.2.5
python-dotenv==1.0.1